+++ src/moduleCommands/ch47f.js
